\name{mean_var_relation}
\alias{mean_var_relation}
\docType{data}
\title{ mean/var relation
}
\description{
  A data frame containing mean/var relation used for simulations. 
}
\usage{data("mean_var_relation")}
\format{
  The format is a list of 6. 
}
\details{
  only elements of x (mean) and y (sqrt of sd) are needed. 
}
\examples{
data(mean_var_relation)
}
\keyword{datasets}
