\name{ICeDT-package}
\alias{ICeDT-package}
\docType{package}
\title{
ICeD-T package
}
\description{
Estimate cell type composition in tumor samples. 
}
\details{
The ICeD-T package is designed to estimate cell type composition from tumor samples while allowing for aberrant gene expression behavior within tumor samples. Please see the manuscript referenced below for further details.
}
\author{
Douglas Roy Wilson, Jr., Chong Jin, and Wei Sun

Maintainer: Wei Sun <wsun@fredhutch.org>
}
\references{
Wilson, Douglas R; Jin, Chong; Ibrahim, Joseph G.; Sun, Wei "ICED-T Provides Accurate Estimates of Immune Cell Abundance in Tumor Samples by Allowing for Aberrant Gene Expression Patterns" <<In Review>>, 2018.
}
\keyword{ package }
\keyword{ Expression Deconvolution }
\keyword{ Tumor }

